Refactors mock webauthn into it's own file and cleans up it's fields
